Rilis v2507.0.0

+++ resources/views/admin/program_bantuan/hasil_pembersihan.blade.php
@@ -22,9 +22,11 @@
                     <div class="box-header with-border">
                         <a href="<?= site_url('program_bantuan') ?>" class="btn btn-social btn-info btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block" title="Kembali Ke Daftar Program Bantuan"><i class="fa fa-arrow-circle-o-left"></i> Kembali Ke Daftar
                             Program Bantuan</a>
-                        <a href="#confirm-delete" title="Hapus Data" onclick="deleteAllBox('mainform','<?= site_url('program_bantuan/bersihkan_data_peserta') ?>')"
-                            class="btn btn-social	btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"
-                        ><i class='fa fa-trash-o'></i> Hapus Data Terpilih</a>
+                        @if (can('h'))
+                            <a href="#confirm-delete" title="Hapus Data" onclick="deleteAllBox('mainform','<?= site_url('program_bantuan/bersihkan_data_peserta') ?>')"
+                                class="btn btn-social	btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"
+                            ><i class='fa fa-trash-o'></i> Hapus Data Terpilih</a>
+                        @endif
                     </div>
                     <div class="box-body">
                         <h4>Data Peserta Tidak Valid</h4>
